+++ terraform/modules/roks_cluster/main.tf
@@ -60,6 +60,7 @@ module "cluster" {
   openshift_cluster_version = var.openshift_cluster_version
   workers_per_zone          = var.roks_workers_per_zone
   min_worker_vcpu_count     = var.roks_min_worker_vcpu_count
+  worker_flavor             = var.roks_worker_flavor
   min_worker_memory_gb      = var.roks_min_worker_memory_gb
   roksbnkctl_binary         = var.roksbnkctl_binary
 }